# generated-api-doc.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- description: Generated server url
  url: http://localhost:8080
...
paths:
  /api/admin/categories:
    get:
      operationId: getCategories
      parameters:
      - in: query
        name: slowdown
        required: false
        schema:
          type: integer
          format: int64
      responses:
        "200":
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CategoryDto"
          description: OK
      tags:
      - Admin
  /api/admin/feedback-dashboard-list:
    get:
      operationId: getFeedbackDashboardList
      parameters:
      - in: query
        name: page
        required: false
        schema:
          type: integer
          format: int32
      - in: query
        name: size
        required: false
        schema:
          type: integer
          format: int32
      - in: query
        name: slowdown
        required: false
        schema:
          type: integer
          format: int64
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageResponseAdminFeedbackDto"
          description: OK
      tags:
      - Admin
  /api/admin/feedback/{feedbackId}:
    delete:
      operationId: deleteFeedback
      parameters:
      - in: path
        name: feedbackId
        required: true
        schema:
          type: integer
          format: int64
      - in: query
        name: slowdown
        required: false
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
      tags:
      - Admin
  /api/admin/feedback/{feedbackId}/approval-status:
    patch:
      operationId: setFeedbackApprovalStatus
      parameters:
      - in: path
        name: feedbackId
        required: true
        schema:
          type: integer
          format: int64
      - in: query
        name: slowdown
        required: false
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetFeedbackApprovalStatusRequest"
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SetFeedbackApprovalStatusResponse"
          description: OK
      tags:
      - Admin
  /api/admin/images:
    get:
      operationId: getAllImages
      parameters:
      - in: query
        name: slowdown
        required: false
        schema:
          type: integer
          format: int64
      responses:
        "200":
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ImageDto"
          description: OK
      tags:
      - Admin
    post:
      operationId: uploadImage
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: Imagefile
                title:
                  type: string
                  description: Title (used for alt attribute)
              required:
              - file
              - title
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ImageDto"
          description: OK
      tags:
      - Admin
  /api/admin/images/{imageId}:
    get:
      operationId: getImageById
      parameters:
      - in: path
        name: imageId
        required: true
        schema:
          type: string
      - in: query
        name: slowdown
        required: false
        schema:
          type: integer
          format: int64
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ImageDto"
          description: OK
      tags:
      - Admin
  /api/admin/me:
    get:
      operationId: getMe
      parameters:
      - in: query
        name: slowdown
        required: false
        schema:
          type: integer
          format: int64
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MeDto"
          description: OK
      tags:
      - Admin
  /api/admin/meal-types:
    get:
      operationId: getMealTypes
      parameters:
      - in: query
        name: slowdown
        required: false
        schema:
          type: integer
          format: int64
      responses:
        "200":
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MealTypeDto"
          description: OK
      tags:
      - Admin
  /api/admin/recipe-dashboard-list:
    get:
      operationId: getRecipeDashboardList
      parameters:
      - in: query
        name: page
        required: false
        schema:
          type: integer
          format: int32
      - in: query
        name: size
        required: false
        schema:
          type: integer
          format: int32
      - in: query
        name: slowdown
        required: false
        schema:
          type: integer
          format: int64
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageResponseAdminRecipeDto"
          description: OK
      tags:
      - Admin
  /api/admin/recipes:
    post:
      operationId: createRecipe
      parameters:
      - in: query
        name: slowdown
        required: false
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecipeRequest"
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AdminRecipeDto"
          description: OK
      tags:
      - Admin
  /api/admin/recipes/{recipeId}:
    get:
      operationId: getRecipe
      parameters:
      - in: path
        name: recipeId
        required: true
        schema:
          type: integer
          format: int64
      - in: query
        name: slowdown
        required: false
        schema:
          type: integer
          format: int64
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AdminRecipeDto"
          description: OK
      tags:
      - Admin
    put:
      operationId: updateRecipe
      parameters:
      - in: path
        name: recipeId
        required: true
        schema:
          type: string
      - in: query
        name: slowdown
        required: false
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecipeRequest"
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AdminRecipeDto"
          description: OK
      tags:
      - Admin
...
  /api/recipes:
    get:
      operationId: recipes
      parameters:
      - in: query
        name: page
        required: false
        schema:
          type: integer
          format: int32
      - in: query
        name: size
        required: false
        schema:
          type: integer
          format: int32
      - in: query
        name: sort
        required: false
        schema:
          type: string
          enum:
          - time
          - likes
          - rating
      - in: query
        name: ids
        required: false
        schema:
          type: string
      - in: query
        name: slowdown
        required: false
        schema:
          type: integer
          format: int64
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageResponseRecipeDto"
          description: OK
      tags:
      - recipe-api-controller
  /api/recipes/{recipeId}:
    get:
      operationId: getRecipe_1
      parameters:
      - in: path
        name: recipeId
        required: true
        schema:
          type: string
      - in: query
        name: slowdown
        required: false
        schema:
          type: integer
          format: int64
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetRecipeResponse"
          description: OK
      tags:
      - recipe-api-controller
...
components:
  schemas:
    AdminFeedbackDto:
      type: object
      properties:
        approvalStatus:
          $ref: "#/components/schemas/ApprovalStatus"
        comment:
          type: string
        commenter:
          type: string
        createdAt:
          type: string
          format: date-time
        id:
          type: string
        rating:
          type: integer
          format: int32
        recipeId:
          type: string
        recipeTitle:
          type: string
        updatedAt:
          type: string
          format: date-time
      required:
      - approvalStatus
      - comment
      - commenter
      - createdAt
      - id
      - rating
      - recipeId
      - recipeTitle
      - updatedAt
    AdminRecipeDto:
      type: object
      properties:
        categoryIds:
          type: array
          items:
            type: string
        cookTime:
          type: integer
          format: int32
        createdAt:
          type: string
          format: date-time
        feedbackCount:
          type: integer
          format: int64
        headline:
          type: string
        id:
          type: string
        image:
          $ref: "#/components/schemas/ImageDto"
        ingredients:
          type: array
          items:
            $ref: "#/components/schemas/IngredientDto"
        instructions:
          type: array
          items:
            $ref: "#/components/schemas/InstructionDto"
        likes:
          type: integer
          format: int32
        mealTypeId:
          type: string
        preparationTime:
          type: integer
          format: int32
        title:
          type: string
        updatedAt:
          type: string
          format: date-time
      required:
      - categoryIds
      - createdAt
      - feedbackCount
      - headline
      - id
      - image
      - ingredients
      - instructions
      - likes
      - mealTypeId
      - title
      - updatedAt
    ApprovalStatus:
      type: string
      enum:
      - APPROVAL_PENDING
      - APPROVED
      - REJECTED
    CategoryDto:
      type: object
      properties:
        description:
          type: string
        icon:
          type: string
        id:
          type: string
        title:
          type: string
        type:
          type: string
      required:
      - description
      - icon
      - id
      - title
      - type
    DetailedRecipeDto:
      type: object
      properties:
        categories:
          type: array
          items:
            $ref: "#/components/schemas/CategoryDto"
        cookTime:
          type: integer
          format: int32
        createdAt:
          type: string
          format: date-time
        headline:
          type: string
        id:
          type: string
        image:
          $ref: "#/components/schemas/ImageDto"
        ingredients:
          type: array
          items:
            $ref: "#/components/schemas/Ingredient"
        instructions:
          type: array
          items:
            $ref: "#/components/schemas/Instruction"
        likes:
          type: integer
          format: int32
        mealType:
          type: string
        preparationTime:
          type: integer
          format: int32
        title:
          type: string
        userFullname:
          type: string
      required:
      - categories
      - cookTime
      - createdAt
      - headline
      - id
      - image
      - ingredients
      - instructions
      - likes
      - mealType
      - preparationTime
      - title
      - userFullname
...
    GetRecipeResponse:
      type: object
      properties:
        recipe:
          $ref: "#/components/schemas/DetailedRecipeDto"
      required:
      - recipe
...
    ImageDto:
      type: object
      properties:
        id:
          type: string
        src:
          type: string
        title:
          type: string
      required:
      - id
      - src
      - title
...
    Ingredient:
      type: object
      properties:
        amount:
          type: number
          format: double
        name:
          type: string
        orderNo:
          type: integer
          format: int32
        unit:
          type: string
      required:
      - amount
      - name
      - orderNo
      - unit
    IngredientDto:
      type: object
      properties:
        amount:
          type: number
          format: double
        name:
          type: string
        unit:
          type: string
      required:
      - amount
      - name
      - unit
    IngredientRequest:
      type: object
      properties:
        amount:
          type: number
          format: double
        name:
          type: string
        unit:
          type: string
      required:
      - amount
      - name
      - unit
    Instruction:
      type: object
      properties:
        description:
          type: string
        orderNo:
          type: integer
          format: int32
    InstructionDto:
      type: object
      properties:
        description:
          type: string
      required:
      - description
    InstructionRequest:
      type: object
      properties:
        description:
          type: string
      required:
      - description
    MeDto:
      type: object
      properties:
        fullname:
          type: string
      required:
      - fullname
    MealTypeDto:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
    NewApprovalStatus:
      type: string
      enum:
      - APPROVED
      - REJECTED
...
    PageResponseAdminFeedbackDto:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/AdminFeedbackDto"
        hasNext:
          type: boolean
        hasPrevious:
          type: boolean
        meta:
          type: string
        pageNumber:
          type: integer
          format: int32
        totalPages:
          type: integer
          format: int32
      required:
      - content
      - hasNext
      - hasPrevious
      - pageNumber
      - totalPages
    PageResponseAdminRecipeDto:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/AdminRecipeDto"
        hasNext:
          type: boolean
        hasPrevious:
          type: boolean
        meta:
          type: string
        pageNumber:
          type: integer
          format: int32
        totalPages:
          type: integer
          format: int32
      required:
      - content
      - hasNext
      - hasPrevious
      - pageNumber
      - totalPages
...
    PageResponseRecipeDto:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/RecipeDto"
        hasNext:
          type: boolean
        hasPrevious:
          type: boolean
        meta:
          type: string
        pageNumber:
          type: integer
          format: int32
        totalPages:
          type: integer
          format: int32
      required:
      - content
      - hasNext
      - hasPrevious
      - pageNumber
      - totalPages
...
    RecipeDto:
      type: object
      properties:
        categories:
          type: array
          items:
            $ref: "#/components/schemas/CategoryDto"
        cookTime:
          type: integer
          format: int32
        createdAt:
          type: string
          format: date-time
        headline:
          type: string
        id:
          type: string
        image:
          $ref: "#/components/schemas/ImageDto"
        likes:
          type: integer
          format: int32
        mealType:
          type: string
        preparationTime:
          type: integer
          format: int32
        title:
          type: string
        userFullname:
          type: string
      required:
      - categories
      - cookTime
      - createdAt
      - headline
      - id
      - image
      - likes
      - mealType
      - preparationTime
      - title
      - userFullname
    RecipeRequest:
      type: object
      properties:
        categoryIds:
          type: array
          items:
            type: string
        cookTime:
          type: integer
          format: int32
        headline:
          type: string
        imageId:
          type: string
        ingredients:
          type: array
          items:
            $ref: "#/components/schemas/IngredientRequest"
        instructions:
          type: array
          items:
            $ref: "#/components/schemas/InstructionRequest"
        mealTypeId:
          type: string
        preparationTime:
          type: integer
          format: int32
        title:
          type: string
      required:
      - categoryIds
      - headline
      - imageId
      - ingredients
      - instructions
      - mealTypeId
      - title
...
    SetFeedbackApprovalStatusRequest:
      type: object
      properties:
        newApprovalStatus:
          $ref: "#/components/schemas/NewApprovalStatus"
      required:
      - newApprovalStatus
    SetFeedbackApprovalStatusResponse:
      type: object
      properties:
        approvalStatus:
          $ref: "#/components/schemas/ApprovalStatus"
      required:
      - approvalStatus